internal restructuring of src/recorder/browser
