+++ src/recorder/browser/index.ts
@@ -1,8 +1,8 @@
-import { startRecording } from './frontend/recording'
-import { initializeView } from './frontend/view'
+import { startRecording } from './recording'
 import { client } from './routing'
 import { configureStorage } from './storage'
 import { isInFrame } from './utils'
+import { initializeView } from './view'
 import { trackTabFocus } from './window'
 
 // CDP will inject this script into all frames, but we only want to run it in the top frame.
