+++ src/recorder/browser/manager.ts
