+++ src/recorder/browser/view/Anchor.tsx
