+++ src/recorder/browser/view/ElementInspector/ElementInspector.hooks.ts
