+++ src/recorder/browser/view/ElementInspector/ElementInspector.tsx
