+++ src/recorder/browser/view/ElementInspector/ElementInspector.utils.ts
