+++ src/recorder/browser/view/ElementInspector/ElementMenu.tsx
