+++ src/recorder/browser/view/ElementInspector/ElementMenu.utils.ts
