+++ src/recorder/browser/view/ElementInspector/ElementPopover.tsx
