+++ src/recorder/browser/view/ElementInspector/assertions/CheckAssertionEditor.tsx
