+++ src/recorder/browser/view/ElementInspector/assertions/VisibilityAssertionEditor.tsx
