+++ src/recorder/browser/view/ElementInspector/hooks.ts
