+++ src/recorder/browser/view/ElementInspector/index.ts
