+++ src/recorder/browser/view/ElementInspector/utils.ts
