+++ src/recorder/browser/view/ElementInspector/waitConditions/types.ts
