+++ src/recorder/browser/view/EventDrawer.tsx
