+++ src/recorder/browser/view/GlobalStyles.tsx
