+++ src/recorder/browser/view/InBrowserControls.tsx
