+++ src/recorder/browser/view/RemoteHighlights.tsx
