+++ src/recorder/browser/view/SettingsProvider.tsx
