+++ src/recorder/browser/view/StudioClientProvider.tsx
