+++ src/recorder/browser/view/TextSelectionPopover.hooks.ts
