+++ src/recorder/browser/view/TextSelectionPopover.tsx
