+++ src/recorder/browser/view/TextSelectionPopover.types.ts
