+++ src/recorder/browser/view/ToolBox/ToolBoxLogo.tsx
