+++ src/recorder/browser/view/ToolBox/ToolBoxRoot.tsx
