+++ src/recorder/browser/view/ToolBox/ToolBoxTooltip.tsx
