+++ src/recorder/browser/view/ToolBox/index.tsx
