+++ src/recorder/browser/view/hooks/useEscape.tsx
