+++ src/recorder/browser/view/hooks/useHighlightDebounce.tsx
