+++ src/recorder/browser/view/hooks/usePreventClick.tsx
