+++ src/recorder/browser/view/hooks/useRecordedEvents.ts
