+++ src/recorder/browser/view/index.tsx
