+++ src/recorder/browser/view/store.ts
