+++ src/recorder/browser/view/types.ts
