+++ src/recorder/browser/view/utils.ts
